backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
tensorflow==2.12.0
Pillow==9.5.0
